+++ tests/EqualityTypes.hs
@@ -14,13 +14,13 @@ module EqualityTypes where
 
 import           Data.Kind
 import           Data.Singletons.TH
-import           Data.Type.Equality ((:~:)(..), (:~~:)(..))
+import           Data.Type.Equality ((:~~:)(..))
 
 import           Internal
 
-data instance Sing :: forall k (a :: k) (b :: k). a :~: b -> Type where
-  SRefl :: Sing Refl
-type (%:~:) = (Sing :: (a :: k) :~: (b :: k) -> Type)
+data (%:~:) :: forall k (a :: k) (b :: k). a :~: b -> Type where
+  SRefl :: (%:~:) Refl
+type instance Sing = (%:~:)
 
 instance SingKind (a :~: b) where
   type Demote (a :~: b) = a :~: b
@@ -47,9 +47,9 @@ instance SingI Refl where
          -> p @@ b
 (~>!:~:) Refl pRefl = pRefl
 
-data instance Sing :: forall j k (a :: j) (b :: k). a :~~: b -> Type where
-  SHRefl :: Sing HRefl
-type (%:~~:) = (Sing :: (a :: j) :~~: (b :: k) -> Type)
+data (%:~~:) :: forall j k (a :: j) (b :: k). a :~~: b -> Type where
+  SHRefl :: (%:~~:) HRefl
+type instance Sing = (%:~~:)
 
 instance SingKind (a :~~: b) where
   type Demote (a :~~: b) = a :~~: b
@@ -78,45 +78,121 @@ instance SingI HRefl where
 
 -----
 
-$(singletons [d|
-  type family WhySym (a :: t) (e :: a :~: (y :: t)) :: Type where
-    WhySym a (_ :: a :~: y) = y :~: a
+-- These newtype wrappers are needed to work around
+-- https://gitlab.haskell.org/ghc/ghc/issues/9269
+newtype WrappedTrans (x :: k) (e :: x :~: y) =
+  WrapTrans { unwrapTrans :: forall (z :: k). y :~: z -> x :~: z }
+newtype WrappedHTrans (x :: j) (e :: x :~~: (y :: k)) =
+  WrapHTrans { unwrapHTrans :: forall l (z :: l). y :~~: z -> x :~~: z }
 
-  type family WhyHsym (a :: j) (e :: a :~~: (y :: z)) :: Type where
-    WhyHsym a (_ :: a :~~: y)  = y :~~: a
+-- This is all needed to avoid impredicativity in the defunctionalization
+-- symbols for WhyHReplace and WhyHLeibniz.
+newtype WrappedPred = WrapPred { unwrapPred :: forall z. z ~> Type }
+type family UnwrapPred (wp :: WrappedPred) :: forall z. z ~> Type where
+  forall (uwp :: forall z. z ~> Type). UnwrapPred (WrapPred uwp) = uwp
+
+$(singletons [d|
+  type WhySym (a :: t) (e :: a :~: (y :: t)) =
+    y :~: a :: Type
+
+  type WhySSym (a :: t) (e :: a :~: (y :: t)) =
+    Sing (Symmetry e) :: Type
+
+  type WhyHSym (a :: j) (e :: a :~~: (y :: z)) =
+    y :~~: a :: Type
+
+  type WhySHSym (a :: j) (e :: a :~~: (y :: z)) =
+    Sing (HSymmetry e) :: Type
 
   type family Symmetry (x :: (a :: k) :~: (b :: k)) :: b :~: a where
     Symmetry Refl = Refl
 
-  type family WhySymIdempotent (a :: t) (r :: a :~: (z :: t)) :: Type where
-    WhySymIdempotent _ r = Symmetry (Symmetry r) :~: r
+  type WhySymIdempotent (a :: t) (r :: a :~: (z :: t)) =
+    Symmetry (Symmetry r) :~: r :: Type
 
-  type family Hsymmetry (x :: a :~~: b) :: b :~~: a where
-    Hsymmetry HRefl = HRefl
+  type family HSymmetry (x :: a :~~: b) :: b :~~: a where
+    HSymmetry HRefl = HRefl
 
-  type family WhyHsymIdempotent (a :: j) (r :: a :~~: (y :: z)) :: Type where
-    WhyHsymIdempotent _ r = Hsymmetry (Hsymmetry r) :~: r
+  type WhyHSymIdempotent (a :: j) (r :: a :~~: (y :: z)) =
+    HSymmetry (HSymmetry r) :~: r :: Type
 
-  type family WhyReplace (from :: t) (p :: t ~> Type)
-                         (e :: from :~: (y :: t)) :: Type where
-    WhyReplace from p (_ :: from :~: y) = p @@ y
+  type WhyTrans (x :: k) (e :: x :~: (y :: k)) =
+    WrappedTrans x e :: Type
 
-  -- Doesn't work due to https://ghc.haskell.org/trac/ghc/ticket/11719
-  {-
-  type family WhyHreplace (from :: j) (p :: forall z. z ~> Type)
-                          (e :: from :~~: (y :: k)) :: Type where
-    WhyHreplace from p (_ :: from :~~: y) = p @@ y
-  -}
+  type WhyHTrans (x :: j) (e :: x :~~: (y :: k)) =
+    WrappedHTrans x e :: Type
 
-  type family WhyLeibniz (f :: t ~> Type) (a :: t) (z :: t) :: Type where
-    WhyLeibniz f a z = f @@ a -> f @@ z
+  type family Trans (x :: a :~: b) (y :: b :~: c) :: a :~: c where
+    Trans Refl Refl = Refl
 
-  type family WhyCong (f :: x ~> y) (a :: x) (e :: a :~: (z :: x)) :: Type where
-    WhyCong (f :: x ~> y) (a :: x) (e :: a :~: (z :: x)) = f @@ a :~: f @@ z
+  type family HTrans (x :: a :~~: b) (y :: b :~~: c) :: a :~~: c where
+    HTrans HRefl HRefl = HRefl
 
-  type family WhyEqIsRefl (a :: k) (e :: a :~: (z :: k)) :: Type where
-    WhyEqIsRefl a e = e :~~: (Refl :: a :~: a)
+  type WhyReplace (from :: t) (p :: t ~> Type) (e :: from :~: (y :: t)) =
+    p @@ y :: Type
 
-  type family WhyHEqIsHRefl (a :: j) (e :: a :~~: (z :: k)) :: Type where
-    WhyHEqIsHRefl a e = e :~~: (HRefl :: a :~~: a)
+  type WhyHReplace (from :: j) (p :: WrappedPred) (e :: from :~~: (y :: k)) =
+    UnwrapPred p @@ y :: Type
+
+  type WhyLeibniz (f :: t ~> Type) (a :: t) (z :: t) =
+    f @@ a -> f @@ z :: Type
+
+  type WhyHLeibniz (f :: WrappedPred) (a :: j) (b :: k) =
+    UnwrapPred f @@ a -> UnwrapPred f @@ b :: Type
+
+  type WhyCong (f :: x ~> y) (a :: x) (e :: a :~: (z :: x)) =
+    f @@ a :~: f @@ z :: Type
+
+  type WhyEqIsRefl (a :: k) (e :: a :~: (z :: k)) =
+    e :~~: (Refl :: a :~: a) :: Type
+
+  type WhyHEqIsHRefl (a :: j) (e :: a :~~: (z :: k)) =
+    e :~~: (HRefl :: a :~~: a) :: Type
+
+  type WhyTransLeft (a :: k) (e :: a :~: (z :: k)) =
+    Trans e Refl :~: e :: Type
+
+  type WhyTransLeftHelper (b :: k) (e :: b :~: (z :: k)) =
+    Trans (Symmetry e) Refl :~: Symmetry e :: Type
+
+  type WhyHTransLeft (a :: j) (e :: a :~~: (z :: k)) =
+    HTrans e HRefl :~: e :: Type
+
+  type WhyHTransLeftHelper (b :: k) (e :: b :~~: (z :: j)) =
+    HTrans (HSymmetry e) HRefl :~: HSymmetry e :: Type
+
+  type WhyTransRight (a :: k) (e :: a :~: (z :: k)) =
+    Trans Refl e :~: e :: Type
+
+  type WhyHTransRight (a :: j) (e :: a :~~: (z :: k)) =
+    HTrans HRefl e :~: e :: Type
+
+  type WhyRebalance (b :: x2 :~: x3) (c :: x3 :~: x4) (a :: x1 :~: x2) =
+    Trans a (Trans b c) :~: Trans (Trans a b) c :: Type
+
+  type WhyRebalanceHelper (b :: x2 :~: x3) (c :: x3 :~: x4) (a :: x2 :~: x1) =
+    Trans (Symmetry a) (Trans b c) :~: Trans (Trans (Symmetry a) b) c :: Type
+
+  type WhyHRebalance (b :: x2 :~~: x3) (c :: x3 :~~: x4) (a :: x1 :~~: x2) =
+    HTrans a (HTrans b c) :~: HTrans (HTrans a b) c :: Type
+
+  type WhyHRebalanceHelper (b :: x2 :~~: x3) (c :: x3 :~~: x4) (a :: x2 :~~: (x1 :: k1)) =
+    HTrans (HSymmetry a) (HTrans b c) :~: HTrans (HTrans (HSymmetry a) b) c :: Type
+  |])
+
+-- These newtype wrappers are needed to work around
+-- https://gitlab.haskell.org/ghc/ghc/issues/9269
+newtype WrappedSTrans (x :: k) (e1 :: x :~: y) =
+  WrapSTrans { unwrapSTrans :: forall (z :: k) (e2 :: y :~: z).
+                               Sing e2 -> Sing (Trans e1 e2) }
+newtype WrappedSHTrans (x :: j) (e1 :: x :~~: (y :: k)) =
+  WrapSHTrans { unwrapSHTrans :: forall l (z :: l) (e2 :: y :~~: z).
+                                 Sing e2 -> Sing (HTrans e1 e2) }
+
+$(singletons [d|
+  type WhySTrans (x :: k) (e :: x :~: (y :: k)) =
+    WrappedSTrans x e :: Type
+
+  type WhySHTrans (x :: j) (e :: x :~~: (y :: k)) =
+    WrappedSHTrans x e :: Type
   |])